args are stored in float[]

+++ modules/renderer/src/com/genir/renderer/bridge/Display.java
@@ -21,7 +21,7 @@ public class Display {
     public static void create(PixelFormat pixel_format) {
         record create(PixelFormat pixel_format) implements GLCommand {
             @Override
-            public void run(Context context, int[] args, int offset) {
+            public void run(Context context, float[] args, int offset) {
                 try {
                     org.lwjgl.opengl.Display.create(pixel_format);
                     context.update();
@@ -41,7 +41,7 @@ public class Display {
     public static void destroy() {
         record destroy() implements GLCommand {
             @Override
-            public void run(Context context, int[] args, int offset) {
+            public void run(Context context, float[] args, int offset) {
                 org.lwjgl.opengl.Display.destroy();
             }
         }
@@ -54,7 +54,7 @@ public class Display {
     public static void update(boolean processMessages) {
         record update(boolean processMessages) implements GLCommand {
             @Override
-            public void run(Context context, int[] args, int offset) {
+            public void run(Context context, float[] args, int offset) {
                 context.update();
                 org.lwjgl.opengl.Display.update(processMessages);
             }
@@ -123,7 +123,7 @@ public class Display {
     public static void setTitle(String newTitle) {
         record setTitle(String newTitle) implements GLCommand {
             @Override
-            public void run(Context context, int[] args, int offset) {
+            public void run(Context context, float[] args, int offset) {
                 org.lwjgl.opengl.Display.setTitle(newTitle);
             }
         }
@@ -134,7 +134,7 @@ public class Display {
     public static void setVSyncEnabled(boolean sync) {
         record setVSyncEnabled(boolean sync) implements GLCommand {
             @Override
-            public void run(Context context, int[] args, int offset) {
+            public void run(Context context, float[] args, int offset) {
                 // Drain OpenGL error queue and log errors. NOTE: As of 0.98a vanilla
                 // ignores OpenGL errors and lets setVSyncEnabled drain the error
                 // queue and throw an exception.
@@ -170,7 +170,7 @@ public class Display {
     public static void setFullscreen(boolean fullscreen) {
         record setFullscreen(boolean fullscreen) implements GLCommand {
             @Override
-            public void run(Context context, int[] args, int offset) {
+            public void run(Context context, float[] args, int offset) {
                 try {
                     org.lwjgl.opengl.Display.setFullscreen(fullscreen);
                     context.update();
@@ -188,7 +188,7 @@ public class Display {
     public static void processMessages() {
         record processMessages() implements GLCommand {
             @Override
-            public void run(Context context, int[] args, int offset) {
+            public void run(Context context, float[] args, int offset) {
                 context.update();
                 org.lwjgl.opengl.Display.processMessages();
             }
@@ -202,7 +202,7 @@ public class Display {
     public static void sync(int fps) {
         record sync(int fps) implements GLCommand {
             @Override
-            public void run(Context context, int[] args, int offset) {
+            public void run(Context context, float[] args, int offset) {
                 org.lwjgl.opengl.Display.sync(fps);
             }
         }
